refactor: move getISODateString
