+++ utils/backend/attendance/getAttendanceOfClass.ts
@@ -1,4 +1,4 @@
-import getISODateString from "@/utils/backend/getISODateString";
+import getISODateString from "@/utils/helpers/getISODateString";
 import logError from "@/utils/helpers/logError";
 import mergeDBLocales from "@/utils/helpers/mergeDBLocales";
 import {
